requirements.txt
nltk==3.8.1
textblob==0.17.1
transformers==4.30.2
torch==2.0.1
numpy==1.24.3
google-cloud-pubsub==2.18.4
